+++ e2e/test/jest-react-ts/verify/verify.ts
@@ -3,10 +3,10 @@ import { expectMetrics } from '../../../helpers';
 describe('After running stryker on jest-react-ts project', () => {
   it('should report expected scores', async () => {
     await expectMetrics({
-      survived: 53,
-      killed: 53,
-      timeout: 6,
-      noCoverage: 0
+      killed: 63,
+      noCoverage: 19,
+      survived: 25,
+      timeout: 5,
     });
   });
 });