+++ packages/jasmine-runner/testResources/big-project/spec/tests/big-text.spec.js
@@ -8,5 +8,5 @@ describe('Big text', () => {
   });
 
   // Leave the garbage collector room to breath
-  afterEach(() => new Promise((res) => setTimeout(res, 10)));
+  afterEach(() => new Promise((res) => setTimeout(res, 100)));
 });