+++ packages/jest-runner/src/jest-test-adapters/jest-less-than-25-adapter.ts
@@ -1,6 +1,4 @@
-import jest from 'jest';
-import { Config } from '@jest/types';
-
+import { jestWrapper } from '../utils';
 import { JestRunResult } from '../jest-run-result';
 
 import { RunSettings, JestTestAdapter } from './jest-test-adapter';
@@ -10,17 +8,20 @@ import { RunSettings, JestTestAdapter } from './jest-test-adapter';
  * It has a lot of `any` typings here, since the installed typings are not in sync.
  */
 export class JestLessThan25TestAdapter implements JestTestAdapter {
-  public run({ jestConfig, projectRoot, fileNameUnderTest, testNamePattern, testLocationInResults }: RunSettings): Promise<JestRunResult> {
+  public run({ jestConfig, fileNameUnderTest, testNamePattern, testLocationInResults }: RunSettings): Promise<JestRunResult> {
     const config = JSON.stringify(jestConfig);
-    return jest.runCLI(
+    return jestWrapper.runCLI(
       {
-        ...(fileNameUnderTest && { _: [fileNameUnderTest], findRelatedTests: true }),
+        $0: 'stryker',
+        _: fileNameUnderTest ? [fileNameUnderTest] : [],
+        findRelatedTests: !!fileNameUnderTest,
         config,
         runInBand: true,
         silent: true,
         testNamePattern,
-      } as Config.Argv,
-      [projectRoot]
+        testLocationInResults,
+      },
+      [jestConfig.rootDir ?? process.cwd()]
     );
   }
 }