+++ packages/jest-runner/src/utils/index.ts
@@ -1,4 +1,3 @@
-export * from './create-react-jest-config';
 export * from './verify-all-test-files-have-coverage';
 export * from './merge-mutant-coverage';
 export * from './jest-wrapper';