+++ packages/jest-runner/test/unit/config-loaders/react-scripts-jest-config-loader.spec.ts
@@ -2,46 +2,33 @@ import path from 'path';
 
 import { expect } from 'chai';
 import sinon from 'sinon';
+import * as utils from '@stryker-mutator/util';
 
 import { ReactScriptsJestConfigLoader } from '../../../src/config-loaders/react-scripts-jest-config-loader';
-import * as helper from '../../../src/utils/create-react-jest-config';
 
 describe(ReactScriptsJestConfigLoader.name, () => {
   let sut: ReactScriptsJestConfigLoader;
   let requireResolveStub: sinon.SinonStub;
+  let resolveStub: sinon.SinonStub;
   let createReactJestConfigStub: sinon.SinonStub;
 
-  const projectRoot = '/path/to/project';
-  const reactScriptsPackagePath = './node_modules/react-scripts/package.json';
-
   beforeEach(() => {
-    createReactJestConfigStub = sinon.stub(helper, 'createReactJestConfig');
-    createReactJestConfigStub.callsFake((resolve: any, root: string, eject: boolean) => ({
-      eject,
-      projectRoot: root,
-      relativePath: resolve('test'),
-    }));
-
-    requireResolveStub = sinon.stub();
-    requireResolveStub.returns(reactScriptsPackagePath);
-
-    sut = new ReactScriptsJestConfigLoader(requireResolveStub as unknown as RequireResolve, projectRoot);
+    requireResolveStub = sinon.stub(utils, 'requireResolve');
+    createReactJestConfigStub = sinon.stub();
+    requireResolveStub.returns(createReactJestConfigStub);
+    createReactJestConfigStub.returns({ testPaths: ['example'] });
+    resolveStub = sinon.stub();
+    resolveStub.returns(path.resolve('./node_modules/react-scripts/package.json'));
+    sut = new ReactScriptsJestConfigLoader(resolveStub as unknown as RequireResolve);
   });
 
   it('should load the configuration via the createJestConfig method provided by react-scripts', () => {
-    sut.loadConfig();
-
-    expect(requireResolveStub).calledWith('react-scripts/package.json');
-  });
-
-  it('should generate a configuration', () => {
     const config = sut.loadConfig();
 
-    expect(config).to.deep.equal({
-      eject: false,
-      projectRoot: '/path/to/project',
-      relativePath: path.join('node_modules', 'react-scripts', 'test'),
-    });
+    expect(resolveStub).calledWith('react-scripts/package.json');
+    expect(requireResolveStub).calledWith('react-scripts/scripts/utils/createJestConfig');
+    expect(createReactJestConfigStub).calledWith(sinon.match.func, process.cwd(), false);
+    expect(config).deep.eq({ testPaths: ['example'] });
   });
 
   it('should throw an error when react-scripts could not be found', () => {
@@ -52,7 +39,7 @@ describe(ReactScriptsJestConfigLoader.name, () => {
 
     // Act & Assert
     expect(() => sut.loadConfig()).throws(
-      'Unable to locate package react-scripts. This package is required when projectType is set to "create-react-app".'
+      'Unable to locate package "react-scripts". This package is required when "jest.projectType" is set to "create-react-app".'
     );
   });
 });