+++ packages/jest-runner/test/unit/jest-test-adapters/jest-greater-than-25-adapter.spec.ts
@@ -11,11 +11,11 @@ describe(JestGreaterThan25TestAdapter.name, () => {
   let sut: JestGreaterThan25TestAdapter;
   let runCLIStub: sinon.SinonStub;
 
-  const projectRoot = '/path/to/project';
   const fileNameUnderTest = '/path/to/file';
-  const jestConfig: Config.InitialOptions = { rootDir: projectRoot };
+  let jestConfig: Config.InitialOptions;
 
   beforeEach(() => {
+    jestConfig = { rootDir: '/path/to/project' };
     runCLIStub = sinon.stub(jestWrapper, 'runCLI');
     runCLIStub.resolves({
       config: jestConfig,
@@ -25,63 +25,69 @@ describe(JestGreaterThan25TestAdapter.name, () => {
     sut = testInjector.injector.injectClass(JestGreaterThan25TestAdapter);
   });
 
-  it('should call the runCLI method with the correct ---projectRoot', async () => {
-    await sut.run({ jestConfig, projectRoot });
-    expect(runCLIStub).calledWith(sinon.match.object, [projectRoot]);
+  it('should call the runCLI method with the correct --projectRoot', async () => {
+    await sut.run({ jestConfig });
+    expect(runCLIStub).calledWith(sinon.match.object, [jestConfig.rootDir!]);
+  });
+
+  it('should call the runCLI method with --projectRoot = cwd when no rootDir is provided', async () => {
+    delete jestConfig.rootDir;
+    await sut.run({ jestConfig });
+    expect(runCLIStub).calledWith(sinon.match.object, [process.cwd()]);
   });
 
   it('should call the runCLI method with the --findRelatedTests flag when provided', async () => {
-    await sut.run({ jestConfig, projectRoot, fileNameUnderTest });
+    await sut.run({ jestConfig, fileNameUnderTest });
 
     expect(runCLIStub).calledWith(
       sinon.match({
         $0: 'stryker',
         _: [fileNameUnderTest],
-        config: JSON.stringify({ rootDir: projectRoot }),
+        config: JSON.stringify(jestConfig),
         findRelatedTests: true,
         runInBand: true,
         silent: true,
         testNamePattern: undefined,
       }),
-      [projectRoot]
+      [jestConfig.rootDir!]
     );
   });
 
   it('should call the runCLI method with the --testNamePattern flag when provided', async () => {
-    await sut.run({ jestConfig, projectRoot, testNamePattern: 'Foo should bar' });
+    await sut.run({ jestConfig, testNamePattern: 'Foo should bar' });
 
     expect(runCLIStub).calledWith(
       sinon.match({
         $0: 'stryker',
         _: [],
-        config: JSON.stringify({ rootDir: projectRoot }),
+        config: JSON.stringify(jestConfig),
         findRelatedTests: false,
         runInBand: true,
         silent: true,
         testNamePattern: 'Foo should bar',
       }),
-      [projectRoot]
+      [jestConfig.rootDir!]
     );
   });
 
   it('should call the runCLI method with the --testLocationInResults flag when provided', async () => {
-    await sut.run({ jestConfig, projectRoot, testLocationInResults: true });
+    await sut.run({ jestConfig, testLocationInResults: true });
 
     expect(runCLIStub).calledWith(
       sinon.match({
         testLocationInResults: true,
       }),
-      [projectRoot]
+      [jestConfig.rootDir!]
     );
   });
 
   it('should call the runCLI method without the --testLocationInResults flag when not', async () => {
-    await sut.run({ jestConfig, projectRoot, testLocationInResults: false });
-    expect(runCLIStub).calledWith(sinon.match({ testLocationInResults: false }), [projectRoot]);
+    await sut.run({ jestConfig, testLocationInResults: false });
+    expect(runCLIStub).calledWith(sinon.match({ testLocationInResults: false }), [jestConfig.rootDir!]);
   });
 
   it('should call the runCLI method and return the test result', async () => {
-    const result = await sut.run({ jestConfig, projectRoot });
+    const result = await sut.run({ jestConfig });
 
     expect(result).to.deep.equal({
       config: jestConfig,
@@ -90,7 +96,7 @@ describe(JestGreaterThan25TestAdapter.name, () => {
   });
 
   it('should call the runCLI method and return the test result when run with --findRelatedTests flag', async () => {
-    const result = await sut.run({ jestConfig, projectRoot, fileNameUnderTest });
+    const result = await sut.run({ jestConfig, fileNameUnderTest });
 
     expect(result).to.deep.equal({
       config: jestConfig,