+++ packages/instrumenter/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.1.0](https://github.com/stryker-mutator/stryker-js/compare/v5.0.1...v5.1.0) (2021-06-14)
+
+**Note:** Version bump only for package @stryker-mutator/instrumenter
+
+
+
+
+
 ## [5.0.1](https://github.com/stryker-mutator/stryker-js/compare/v5.0.0...v5.0.1) (2021-05-28)
 
 **Note:** Version bump only for package @stryker-mutator/instrumenter